refactor: Extract contracts and logging
