+++ mintbase-deps/src/lib.rs
@@ -5,15 +5,21 @@ mod impls;
 mod structs;
 mod traits;
 
-use std::path::PathBuf;
-use std::str::FromStr;
-
 pub use consts::*;
 pub use enums::*;
 pub use fns::*;
 pub use impls::*;
 pub use structs::*;
 pub use traits::*;
+
+#[cfg(feature = "factory-wasm")]
+pub mod factory;
+#[cfg(feature = "helper-wasm")]
+pub mod helper;
+#[cfg(feature = "store-wasm")]
+pub mod store;
+
+pub mod logging;
 
 #[cfg(feature = "all")]
 mod mintbase_std {
@@ -44,6 +50,28 @@ mod mintbase_std {
         uuid,
     };
 }
+#[cfg(feature = "all")]
+pub use logging::{
+    Nep171EventLog,
+    NftApproveLog,
+    NftBurnLog,
+    NftComposeLog,
+    NftListLog,
+    NftLoanSetLog,
+    NftMarketLog,
+    NftMintLog,
+    NftMintLogMemo,
+    NftOfferLog2,
+    NftOptionStringLog,
+    NftRevokeLog,
+    NftSaleLog,
+    NftSetSplitOwnerLog,
+    NftStoreCreateLog,
+    NftStringLog,
+    NftTransferLog,
+    NftUpdateListLog,
+    NftUpdateOfferLog,
+};
 #[cfg(feature = "all")]
 pub use mintbase_std::*;
 #[cfg(feature = "wasm")]
