+++ mintbase-deps/src/structs.rs
@@ -3,7 +3,6 @@ use std::collections::HashMap;
 use near_sdk::json_types::{
     Base64VecU8,
     U128,
-    U64,
 };
 use near_sdk::{
     AccountId,
@@ -301,96 +300,10 @@ pub struct StoreInitArgs {
     pub owner_id: AccountId,
 }
 
-#[cfg_attr(
-    feature = "helper-wasm",
-    near_sdk::near_bindgen,
-    derive(BorshDeserialize, BorshSerialize)
-)]
-#[cfg(feature = "helper-wasm")]
-pub struct HelperWasm {
-    pub count: u64,
-}
-
-#[cfg_attr(
-    feature = "store-wasm",
-    near_sdk::near_bindgen,
-    derive(BorshDeserialize, BorshSerialize)
-)]
-#[cfg(feature = "store-wasm")]
-pub struct MintbaseStore {
-    /// Accounts that are allowed to mint tokens on this Store.
-    pub minters: UnorderedSet<AccountId>,
-    /// Initial deployment data of this Store.
-    pub metadata: NFTContractMetadata,
-    /// If a Minter mints more than one token at a time, all tokens will
-    /// share the same `TokenMetadata`. It's more storage-efficient to store
-    /// that `TokenMetadata` once, rather than to copy the data on each
-    /// Token. The key is generated from `tokens_minted`. The map keeps count
-    /// of how many copies of this token remain, so that the element may be
-    /// dropped when the number reaches zero (ie, when tokens are burnt).
-    pub token_metadata: LookupMap<u64, (u16, TokenMetadata)>,
-    /// If a Minter mints more than one token at a time, all tokens will
-    /// share the same `Royalty`. It's more storage-efficient to store that
-    /// `Royalty` once, rather than to copy the data on each Token. The key
-    /// is generated from `tokens_minted`. The map keeps count of how many
-    /// copies of this token remain, so that the element may be dropped when
-    /// the number reaches zero (ie, when tokens are burnt).
-    pub token_royalty: LookupMap<u64, (u16, Royalty)>,
-    /// Tokens this Store has minted, excluding those that have been burned.
-    pub tokens: LookupMap<u64, Token>,
-    /// A mapping from each user to the tokens owned by that user. The owner
-    /// of the token is also stored on the token itself.
-    pub tokens_per_owner: LookupMap<AccountId, UnorderedSet<u64>>,
-    /// A map from a token_id of a token on THIS contract to a set of tokens,
-    /// that may be on ANY contract. If the owned-token is on this contract,
-    /// the id will have format "<u64>". If the token is on another contract,
-    /// the token will have format "<u64>:account_id"
-    pub composeables: LookupMap<String, UnorderedSet<String>>,
-    /// The number of tokens this `Store` has minted. Used to generate
-    /// `TokenId`s.
-    pub tokens_minted: u64,
-    /// The number of tokens this `Store` has burned.
-    pub tokens_burned: u64,
-    /// The number of tokens approved (listed) by this `Store`. Used to index
-    /// listings and approvals. List ID format: `list_nonce:token_key`
-    pub num_approved: u64,
-    /// The owner of the Contract.
-    pub owner_id: AccountId,
-    /// The Near-denominated price-per-byte of storage, and associated
-    /// contract storage costs. As of April 2021, the price per bytes is set
-    /// to 10^19, but this may change in the future, thus this
-    /// future-proofing field.
-    pub storage_costs: StorageCosts,
-    /// If false, disallow users to call `nft_move`.
-    pub allow_moves: bool,
-}
-
 #[cfg(feature = "all")]
 pub struct MyMakeWriter {
     pub stdout: std::io::Stdout,
     pub stderr: std::io::Stderr,
-}
-
-#[cfg(feature = "factory-wasm")]
-#[near_bindgen]
-#[cfg(feature = "factory-wasm")]
-#[derive(BorshDeserialize, BorshSerialize)]
-pub struct MintbaseStoreFactory {
-    /// The `Store`s this `Factory` has produced.
-    pub stores: LookupSet<String>,
-    /// Fee taken by Mintbase for `Store` deployment.
-    pub mintbase_fee: Balance,
-    /// The owner may update the `mintbase_fee`.
-    pub owner_id: AccountId,
-    /// The Near-denominated price-per-byte of storage. As of April 2021, the
-    /// price per bytes is set by default to 10^19, but this may change in the
-    /// future, thus this future-proofing field.
-    pub storage_price_per_byte: u128,
-    /// Cost in yoctoNear to deploy a store. Changes if `storage_price_per_byte`
-    /// changes.
-    pub store_cost: u128,
-    /// The public key to give a full access key to
-    pub admin_public_key: PublicKey,
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -412,207 +325,6 @@ pub struct NFTContractMetadata {
     /// Base64-encoded sha256 hash of the JSON file pointed at by the reference
     /// field. Required if `reference` is included.
     pub reference_hash: Option<Base64VecU8>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Nep171Event {
-    pub standard: String,
-    pub version: String,
-    #[serde(flatten)]
-    pub event_kind: Nep171EventLog,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NearJsonEvent {
-    pub standard: String,
-    pub version: String,
-    pub event: String,
-    pub data: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftStoreCreateLog {
-    pub contract_metadata: NFTContractMetadata,
-    pub owner_id: String,
-    pub id: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftStringLog {
-    pub data: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftOptionStringLog {
-    pub data: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftMintLog {
-    pub owner_id: String,
-    pub token_ids: Vec<String>,
-    pub memo: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftBurnLog {
-    pub owner_id: String,
-    pub authorized_id: Option<String>,
-    pub token_ids: Vec<String>,
-    pub memo: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftApproveLog {
-    pub token_id: u64,
-    pub approval_id: u64,
-    pub account_id: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftRevokeLog {
-    pub token_id: u64,
-    pub account_id: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftTransferLog {
-    pub authorized_id: Option<String>,
-    pub old_owner_id: String,
-    pub new_owner_id: String,
-    pub token_ids: Vec<String>,
-    pub memo: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftSetSplitOwnerLog {
-    pub split_owners: SplitOwners,
-    pub token_ids: Vec<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftLoanSetLog {
-    pub account_id: Option<String>,
-    pub token_id: u64,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftComposeLog {
-    pub token_ids: Vec<U64>,
-    /// direct parent of token_ids
-    pub parent: String,
-    /// - "t": owned directly by a token on this contract
-    /// - "k": owned directly by a token on another contract
-    pub ttype: String,
-    /// local root of chain of token_ids
-    pub lroot: Option<u64>,
-    /// holder of local root
-    pub holder: String,
-    pub depth: u8,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftUncomposeLog {
-    pub token_ids: Vec<U64>,
-    pub holder: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftOnComposeLog {
-    pub predecessor: String,
-    pub token_id: U64,
-    /// direct parent of token_ids
-    pub cross_child_id: U64,
-    /// local root of chain of token_ids
-    pub lroot: Option<u64>,
-    /// holder of local root
-    pub holder: String,
-    pub depth: u8,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftOnUncomposeLog {
-    pub token_id: U64,
-    pub holder: String,
-    pub child_key: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftMovedLog {
-    pub token_id: U64,
-    pub contract_id: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftOnMoveLog {
-    pub token_id: U64,
-    pub origin_key: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftListLog {
-    pub list_id: String,
-    pub price: String,
-    pub token_key: String,
-    pub owner_id: String,
-    pub autotransfer: bool,
-    pub approval_id: String,
-    pub token_id: String,
-    pub store_id: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftMintLogMemo {
-    pub royalty: Option<Royalty>,
-    pub split_owners: Option<SplitOwners>,
-    pub meta_id: Option<String>,
-    pub meta_extra: Option<String>,
-    pub minter: String,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftUpdateListLog {
-    pub auto_transfer: Option<bool>,
-    pub price: Option<String>,
-    pub list_id: Option<String>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftOfferLog2 {
-    pub offer: TokenOffer,
-    pub list_id: String,
-    pub token_key: String,
-    pub offer_num: u64,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftOfferLog {
-    pub price: String,
-    pub from: String,
-    pub timeout: String,
-    pub list_id: String,
-    pub token_key: String,
-    pub offer_num: u64,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftUpdateOfferLog {
-    pub list_id: String,
-    pub offer_num: u64,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftSaleLog {
-    pub list_id: String,
-    pub offer_num: u64,
-    pub token_key: String,
-    pub payout: HashMap<AccountId, U128>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NftMarketLog {
-    pub account_id: String,
-    pub state: bool,
 }
 
 #[derive(Debug, Clone)]
