+++ mintbase-deps/src/traits.rs
@@ -5,19 +5,6 @@ use near_sdk::*;
 
 use crate::*;
 
-#[cfg_attr(feature = "factory-wasm", ext_contract(factory_self))]
-#[cfg(feature = "factory-wasm")]
-pub trait OnCreateCallback {
-    fn on_create(
-        &mut self,
-        store_creator_id: AccountId,
-        metadata: NFTContractMetadata,
-        owner_id: AccountId,
-        store_account_id: AccountId,
-        attached_deposit: U128,
-    );
-}
-
 pub trait New {
     fn new(arg: Self) -> Self;
 }
@@ -146,46 +133,6 @@ pub trait NonFungibleContractCore {
         &self,
         token_id: U64,
     ) -> Option<Token>;
-}
-
-#[cfg_attr(feature = "store-wasm", ext_contract(store_self))]
-#[cfg(feature = "store-wasm")]
-pub trait NonFungibleResolveTransfer {
-    /// Finalize an `nft_transfer_call` chain of cross-contract calls.
-    ///
-    /// The `nft_transfer_call` process:
-    ///
-    /// 1. Sender calls `nft_transfer_call` on FT contract
-    /// 2. NFT contract transfers token from sender to receiver
-    /// 3. NFT contract calls `nft_on_transfer` on receiver contract
-    /// 4+. [receiver contract may make other cross-contract calls]
-    /// N. NFT contract resolves promise chain with `nft_resolve_transfer`, and may
-    ///    transfer token back to sender
-    ///
-    /// Requirements:
-    /// * Contract MUST forbid calls to this function by any account except self
-    /// * If promise chain failed, contract MUST revert token transfer
-    /// * If promise chain resolves with `true`, contract MUST return token to
-    ///   `sender_id`
-    ///
-    /// Arguments:
-    /// * `sender_id`: the sender of `ft_transfer_call`
-    /// * `token_id`: the `token_id` argument given to `ft_transfer_call`
-    /// * `approved_token_ids`: if using Approval Management, contract MUST provide
-    ///   set of original approved accounts in this argument, and restore these
-    ///   approved accounts in case of revert.
-    ///
-    /// Returns true if token was successfully transferred to `receiver_id`.
-    ///
-    /// Mild modifications from core standard, commented where applicable.
-    #[private]
-    fn nft_resolve_transfer(
-        &mut self,
-        owner_id: AccountId,
-        receiver_id: AccountId,
-        token_id: String,
-        approved_account_ids: Option<Vec<String>>,
-    );
 }
 
 /// Non-Fungible Token Approval NEP 178. Ref:
